removes param_grid as parameter. Hard codes for MVP
